+++ src/views/ProjectPage/components/common/EmptyContent.tsx
@@ -1,6 +1,6 @@
 import cc from 'classcat';
 
-import styles from '../styles/empty.module.scss';
+import styles from './empty.module.scss';
 import { ReactComponent as Circle } from '@src/assets/icons/circle.svg';
 import { ReactComponent as Diagonal } from '@src/assets/icons/diagonal.svg';
 import { ReactComponent as Triangle } from '@src/assets/icons/triangle.svg';
