+++ src/views/ProjectPage/components/filter/MobileFilter.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { MobileFilterModal } from './MobileFilterModal';
 import { MobileUtilityButtons } from './MobileUtilityButtons';
-import { ProjectCategoryType } from '../lib/constants';
+import { ProjectCategoryType } from '../../lib/constants';
 import { Condition } from '@src/lib';
 
 type MobileFilterProps = {
