+++ src/views/ProjectPage/components/filter/index.ts
@@ -1,0 +1,2 @@
+export * from './DesktopFilter';
+export * from './MobileFilter';
