+++ src/views/ProjectPage/components/index.ts
@@ -1,6 +1,6 @@
-export * from './DesktopFilter';
-export * from './EmptyContent';
-export * from './MobileFilter';
-export * from './ProjectCard';
-export * from './ProjectEnrollSection';
-export * from './ProjectList';
+export * from './common/EmptyContent';
+export * from './filter/DesktopFilter';
+export * from './filter/MobileFilter';
+export * from './project/ProjectCard';
+export * from './project/ProjectEnrollSection';
+export * from './project/ProjectList';
