+++ src/views/ProjectPage/components/project/ProjectCard.tsx
@@ -2,15 +2,15 @@ import Image from 'next/image';
 import Link from 'next/link';
 import shortid from 'shortid';
 
-import styles from '../styles/project-card.module.scss';
-import { ProjectType, LinkType } from '../types';
+import styles from './project-card.module.scss';
+import { ProjectType, LinkType } from '../../types';
 
-import { ReactComponent as PlayStore } from '../assets/play-store-40x40.svg';
-import { ReactComponent as AppStore } from '../assets/app-store-40x40.svg';
-import { ReactComponent as Github } from '../assets/github-40x40.svg';
-import { ReactComponent as Web } from '../assets/website-40x40.svg';
-import { ReactComponent as Instagram } from '../assets/instagram-30x30.svg';
-import { ReactComponent as Youtube } from '../assets/youtube-30x30.svg';
+import { ReactComponent as PlayStore } from '../../assets/play-store-40x40.svg';
+import { ReactComponent as AppStore } from '../../assets/app-store-40x40.svg';
+import { ReactComponent as Github } from '../../assets/github-40x40.svg';
+import { ReactComponent as Web } from '../../assets/website-40x40.svg';
+import { ReactComponent as Instagram } from '../../assets/instagram-30x30.svg';
+import { ReactComponent as Youtube } from '../../assets/youtube-30x30.svg';
 
 export function ProjectCard({ project }: { project: ProjectType }) {
   return (
