+++ src/views/ProjectPage/components/project/project-card.module.scss
